+++ apps/omsorgspengesoknad-v2/src/app/steps/delt-bosted/i18n/nn.ts
@@ -4,4 +4,8 @@ export const deltBostedStegMessages_nn: Record<keyof typeof deltBostedStegMessag
     'deltBostedSteg.counsellorpanel':
         'For å få ekstra omsorgsdagar må du bu saman med barnet. Du har opplyst om at de har avtale om delt fast bustad. Viss barnet ikkje er folkeregistrert på adressa di, må du senda inn avtalen om delt fast bustad. Du lastar opp avtalen anten ved å ta bilete av han, eller ved å skanna han.',
     'deltBostedSteg.samværsavtale.label': 'Last opp avtalen',
+    'deltBostedForm.validation.samværsavtale.noVedleggUploaded': 'Ingen dokument er lasta opp',
+    'deltBostedForm.validation.samværsavtale.tooManyVedlegg': 'For mange dokument er lasta opp',
+    'deltBostedForm.validation.samværsavtale.maxTotalSizeExceeded':
+        'Total samla storleik for dokumenta du har lasta opp overstig grensa på 24 MB.',
 };