+++ apps/opplaringspenger-soknad/src/app/søknad/steps/legeerklæring/legeerklæringMessages.ts
@@ -14,7 +14,7 @@ const nb = {
     'vedleggsliste.ingenLegeerklæringLastetOpp': 'Ingen dokumenter er lastet opp',
 
     'dokumenter.advarsel.totalstørrelse.1':
-        'Du har totalt lastet opp mer enn grensen på 24 MB. Det betyr at du må fjerne noe av det du har lastet opp. Hvis det betyr at du ikke får plass til alt du ønsker å sende nå, kan du ',
+        'Du har totalt lastet opp mer enn grensen på 24 MB. Det betyr at du må fjerne noe av det du har lastet opp. Hvis du da ikke får plass til alt du ønsker å sende nå, kan du ',
     'dokumenter.advarsel.totalstørrelse.2': 'ettersende flere dokumenter.',
 
     'validation.skalEttersendeVedlegg.yesOrNoIsUnanswered': 'Du må svare på om du skal ettersende vedlegg',