+++ packages/sif-soknad-ui/src/pages/start-page/StartPage.tsx
@@ -1,8 +1,20 @@
-import { BodyLong, Box, Button, Checkbox, CheckboxGroup, GuidePanel, Heading, Link, VStack } from '@navikt/ds-react';
-import { ReactNode, useState } from 'react';
+import {
+    BodyLong,
+    Box,
+    Button,
+    Checkbox,
+    CheckboxGroup,
+    ErrorSummary,
+    GuidePanel,
+    Heading,
+    Link,
+    VStack,
+} from '@navikt/ds-react';
+import { ReactNode, SubmitEvent, useRef, useState } from 'react';
 
 import { SifSoknadUiText, useSifSoknadUiIntl } from '../../i18n';
 import { ApplicationPage } from '../application-page/ApplicationPage';
+import { ArrowRightIcon } from '@navikt/aksel-icons';
 
 interface Props {
     /** Tittel på søknad/app */
@@ -24,16 +36,19 @@ export const StartPage = ({ title, guide, children, onStart, isPending }: Props)
     const { text } = useSifSoknadUiIntl();
     const [error, setError] = useState(false);
     const [bekrefter, setBekrefter] = useState(false);
+    const summaryRef = useRef<HTMLDivElement>(null);
 
     const isValid = () => {
         if (!bekrefter) {
             setError(true);
+            setTimeout(() => summaryRef.current?.focus(), 100);
             return false;
         }
         setError(false);
         return true;
     };
-    const handleSubmit = (event: React.FormEvent) => {
+
+    const handleSubmit = (event: SubmitEvent) => {
         event.preventDefault();
         if (isValid()) {
             onStart(true);
@@ -55,36 +70,56 @@ export const StartPage = ({ title, guide, children, onStart, isPending }: Props)
                 <div>{children}</div>
                 <section aria-label={text('@sifSoknadUi.startPage.form.ariaLabel')}>
                     <form onSubmit={handleSubmit}>
-                        <div>
-                            <BodyLong spacing>
+                        <VStack gap="space-24">
+                            <BodyLong>
                                 <SifSoknadUiText id="@sifSoknadUi.startPage.disclosure.text" />{' '}
                                 <Link href="https://www.nav.no/endringer">
                                     <SifSoknadUiText id="@sifSoknadUi.startPage.disclosure.linkText" />
                                 </Link>
                                 .
                             </BodyLong>
-                            <Box paddingBlock="space-0 space-32">
-                                <CheckboxGroup
-                                    legend={text('@sifSoknadUi.startPage.confirmation.legend')}
-                                    hideLegend={true}
-                                    error={error ? text('@sifSoknadUi.startPage.confirmation.error') : undefined}>
-                                    <Checkbox
-                                        name="bekrefter"
-                                        value="bekrefter"
-                                        onChange={(evt) => {
-                                            setBekrefter(evt.target.checked);
-                                            if (evt.target.checked) {
-                                                setError(false);
-                                            }
-                                        }}>
-                                        <SifSoknadUiText id="@sifSoknadUi.startPage.confirmation.checkboxLabel" />
-                                    </Checkbox>
-                                </CheckboxGroup>
-                            </Box>
-                            <Button type="submit" loading={isPending} disabled={isPending}>
+
+                            <CheckboxGroup
+                                legend={text('@sifSoknadUi.startPage.confirmation.legend')}
+                                hideLegend={true}
+                                error={error ? text('@sifSoknadUi.startPage.confirmation.error') : undefined}>
+                                <Checkbox
+                                    id="bekrefter"
+                                    name="bekrefter"
+                                    value="bekrefter"
+                                    onChange={(evt) => {
+                                        setBekrefter(evt.target.checked);
+                                        if (evt.target.checked) {
+                                            setError(false);
+                                        }
+                                    }}>
+                                    <SifSoknadUiText id="@sifSoknadUi.startPage.confirmation.checkboxLabel" />
+                                </Checkbox>
+                            </CheckboxGroup>
+
+                            {error && (
+                                <div ref={summaryRef} tabIndex={-1}>
+                                    <ErrorSummary heading={text('@sifSoknadUi.startPage.errorSummary.heading')}>
+                                        <ErrorSummary.Item
+                                            href="#bekrefter"
+                                            onClick={(evt) => {
+                                                evt.preventDefault();
+                                                document.getElementsByName('bekrefter')[0]?.focus();
+                                            }}>
+                                            {text('@sifSoknadUi.startPage.confirmation.error')}
+                                        </ErrorSummary.Item>
+                                    </ErrorSummary>
+                                </div>
+                            )}
+                            <Button
+                                type="submit"
+                                loading={isPending}
+                                disabled={isPending}
+                                icon={<ArrowRightIcon />}
+                                iconPosition="right">
                                 <SifSoknadUiText id="@sifSoknadUi.startPage.submitButton" />
                             </Button>
-                        </div>
+                        </VStack>
                     </form>
                 </section>
             </VStack>