+++ packages/sif-soknad-ui/src/pages/start-page/i18n/nn.ts
@@ -1,5 +1,14 @@
 import { startPageMessages_nb } from './nb';
 
 export const startPageMessages_nn: Record<keyof typeof startPageMessages_nb, string> = {
-    ...startPageMessages_nb,
+    '@sifSoknadUi.startPage.guide.greeting': 'Hei {navn}',
+    '@sifSoknadUi.startPage.form.ariaLabel': 'Skjema',
+    '@sifSoknadUi.startPage.disclosure.text':
+        'Det er viktig at du gir oss rette opplysningar slik at me kan handsame saka di.',
+    '@sifSoknadUi.startPage.disclosure.linkText': 'Les meir om kor viktig det er å gje rette opplysningar',
+    '@sifSoknadUi.startPage.confirmation.legend': 'Eg stadfestar at eg vil svare så rett som eg kan',
+    '@sifSoknadUi.startPage.confirmation.error': 'Du må stadfeste at du vil svare så rett som du kan',
+    '@sifSoknadUi.startPage.confirmation.checkboxLabel': 'Eg stadfestar at eg vil svare så rett som eg kan.',
+    '@sifSoknadUi.startPage.submitButton': 'Start søknad',
+    '@sifSoknadUi.startPage.errorSummary.heading': 'Du må rette opp følgjande feil',
 };